Supporting material swapping and class documentation.
