+++ Assets/MixedRealityToolkit/Interfaces/Rendering.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5b45180758e7f94e88ba59cdfb25833
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
