+++ Assets/MixedRealityToolkit/Interfaces/Rendering/IMaterialInstanceOwner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 798ce43767beef14db83ef09f89528d0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
