+++ Assets/MixedRealityToolkit/Utilities/Rendering.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 22471e245140aec4a9b57f84b7f1cb8c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
