+++ Assets/MixedRealityToolkit/Utilities/Rendering/MaterialInstance.cs
@@ -1,0 +1,303 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Collections.Generic;
+using UnityEditor;
+
+#if UNITY_EDITOR
+using UnityEngine;
+#endif
+
+namespace Microsoft.MixedReality.Toolkit.Rendering
+{
+    /// <summary>
+    /// The MaterialInstance behavior aides in tracking instance material lifetime(s) and automatically destroys instanced materials. 
+    /// This utility component can be uses as a replacement to <see href="https://docs.unity3d.com/ScriptReference/Renderer-material.html">Renderer.material</see> or 
+    /// <see href="https://docs.unity3d.com/ScriptReference/Renderer-materials.html">Renderer.materials</see>. When invoking Unity's Renderer.material(s) Unity 
+    /// automatically instantiates new materials. It is the caller's responsibility to destroy the materials when the game object is destroyed. The MaterialInstance helps 
+    /// avoid material leaks. The MaterialInstance behavior also keeps material allocation paths consistent for during edit time and run time.
+    /// </summary>
+    [ExecuteAlways, RequireComponent(typeof(Renderer))]
+    public class MaterialInstance : MonoBehaviour
+    {
+        /// <summary>
+        /// Returns the first instantiated Material assigned to the renderer, similar to <see href="https://docs.unity3d.com/ScriptReference/Renderer-material.html">Renderer.material</see>. 
+        /// If any owner is specified the instanced material(s) will not be released until all owners are released. When a material
+        /// is no longer needed ReleaseMaterial should be called with the matching owner.
+        /// </summary>
+        /// <param name="owner">An optional owner to track instance ownership.</param>
+        /// <returns>The first instantiated Material.</returns>
+        public Material AcquireMaterial(object owner)
+        {
+            if (owner != null)
+            {
+                materialOwners.Add(owner);
+            }
+
+            return Material;
+        }
+
+        /// <summary>
+        /// Returns all the instantiated materials of this object, similar to <see href="https://docs.unity3d.com/ScriptReference/Renderer-materials.html">Renderer.materials</see>. 
+        /// If any owner is specified the instanced material(s) will not be released until all owners are released. When a material
+        /// is no longer needed ReleaseMaterial should be called with the matching owner.
+        /// </summary>
+        /// <param name="owner">An optional owner to track instance ownership.</param>
+        /// <returns>All the instantiated materials.</returns>
+        public Material[] AcquireMaterials(object owner)
+        {
+            if (owner != null)
+            {
+                materialOwners.Add(owner);
+            }
+
+            return Materials;
+        }
+
+        /// <summary>
+        /// Relinquishes ownership of a material instance. This should be called when a material is no longer needed
+        /// after acquire ownership with AcquireMaterial(s).
+        /// </summary>
+        /// <param name="owner">The same owner which originally acquire ownership via AcquireMaterial(s).</param>
+        /// <param name="autoDestroy">When ownership count hits zero should the MaterialInstance component be destroyed?</param>
+        public void ReleaseMaterial(object owner, bool autoDestroy = true)
+        {
+            materialOwners.Remove(owner);
+
+            if (autoDestroy && materialOwners.Count == 0)
+            {
+                DestorySafe(this);
+            }
+        }
+
+        /// <summary>
+        /// Returns the first instantiated Material assigned to the renderer, similar to <see href="https://docs.unity3d.com/ScriptReference/Renderer-material.html">Renderer.material</see>.
+        /// </summary>
+        public Material Material
+        {
+            get
+            {
+                AcquireInstances();
+
+                if (instanceMaterials?.Length > 0)
+                {
+                    return instanceMaterials[0];
+                }
+
+                return null;
+            }
+        }
+
+        /// <summary>
+        /// Returns all the instantiated materials of this object, similar to <see href="https://docs.unity3d.com/ScriptReference/Renderer-materials.html">Renderer.materials</see>.
+        /// </summary>
+        public Material[] Materials
+        {
+            get
+            {
+                AcquireInstances();
+
+                return instanceMaterials;
+            }
+        }
+
+        private Renderer CachedRenderer
+        {
+            get
+            {
+                if (cachedRenderer == null)
+                {
+                    cachedRenderer = GetComponent<Renderer>();
+                }
+
+                return cachedRenderer;
+            }
+        }
+
+        private Renderer cachedRenderer = null;
+
+        [SerializeField, HideInInspector]
+        private Material[] defaultMaterials = null;
+        private Material[] instanceMaterials = null;
+        private bool materialsInstanced = false;
+        private HashSet<object> materialOwners = new HashSet<object>();
+
+        private const string instancePostfix = " (Clone)";
+
+#region MonoBehaviour Implementation
+
+        private void Awake()
+        {
+            if (CachedRenderer != null)
+            {
+                // Cache the default materials if ones do not already exist.
+                if (defaultMaterials == null || defaultMaterials.Length == 0)
+                {
+                    defaultMaterials = CachedRenderer.sharedMaterials;
+                }
+                else if (!materialsInstanced) // Restore the clone to it's initial state.
+                {
+                    CachedRenderer.sharedMaterials = defaultMaterials;
+                }
+            }
+        }
+
+        private void Update()
+        {
+            // If the materials get changed via outside of MaterialInstance.
+            var sharedMaterials = CachedRenderer.sharedMaterials;
+
+            if (!MaterialsMatch(sharedMaterials, instanceMaterials))
+            {
+                // Re-create the material instances.
+                var newDefaultMaterials = new Material[sharedMaterials.Length];
+                var min = Math.Min(newDefaultMaterials.Length, defaultMaterials.Length);
+
+                // Copy the old defaults.
+                for (var i = 0; i < min; ++i)
+                {
+                    newDefaultMaterials[i] = defaultMaterials[i];
+                }
+
+                // Patch in the new defaults.
+                for (var i = 0; i < newDefaultMaterials.Length; ++i)
+                {
+                    var material = sharedMaterials[i];
+
+                    if (!IsInstanceMaterial(material))
+                    {
+                        newDefaultMaterials[i] = material;
+                    }
+                }
+
+                defaultMaterials = newDefaultMaterials;
+                CreateInstances();
+
+                // Notify owners of the change.
+                foreach (var owner in materialOwners)
+                {
+                    var materialInstanceOwner = owner as IMaterialInstanceOwner;
+
+                    materialInstanceOwner?.OnMaterialChanged(this);
+                }
+            }
+        }
+
+        private void OnDestroy()
+        {
+            if (CachedRenderer != null)
+            {
+                CachedRenderer.sharedMaterials = defaultMaterials;
+            }
+
+            DestroyMaterials(instanceMaterials);
+        }
+
+#endregion MonoBehaviour Implementation
+
+        private void AcquireInstances()
+        {
+            if (CachedRenderer != null)
+            {
+                if (!MaterialsMatch(CachedRenderer.sharedMaterials, instanceMaterials))
+                {
+                    CreateInstances();
+                }
+            }
+        }
+
+        private void CreateInstances()
+        {
+            DestroyMaterials(instanceMaterials);
+            instanceMaterials = InstanceMaterials(defaultMaterials);
+
+            if (CachedRenderer != null)
+            {
+                CachedRenderer.sharedMaterials = instanceMaterials;
+            }
+
+            materialsInstanced = true;
+        }
+
+        private static bool MaterialsMatch(Material[] a, Material[] b)
+        {
+            if (a?.Length != b?.Length)
+            {
+                return false;
+            }
+
+            for (int i = 0; i < a?.Length; ++i)
+            {
+                if (a[i] != b[i])
+                {
+                    return false;
+                }
+            }
+
+            return true;
+        }
+
+        private static Material[] InstanceMaterials(Material[] source)
+        {
+            if (source == null)
+            {
+                return null;
+            }
+
+            var output = new Material[source.Length];
+
+            for (var i = 0; i < source.Length; ++i)
+            {
+                if (source[i] != null)
+                {
+                    Debug.Assert(!IsInstanceMaterial(source[i]), "A material which is already instanced was instanced multiple times.");
+
+                    output[i] = new Material(source[i]);
+                    output[i].name += instancePostfix;
+                }
+            }
+
+            return output;
+        }
+
+        private static void DestroyMaterials(Material[] materials)
+        {
+            if (materials != null)
+            {
+                for (var i = 0; i < materials.Length; ++i)
+                {
+                    DestorySafe(materials[i]);
+                }
+
+                Array.Clear(materials, 0, materials.Length);
+            }
+        }
+
+        private static bool IsInstanceMaterial(Material material)
+        {
+            return ((material != null) && material.name.Contains(instancePostfix));
+        }
+
+        private static void DestorySafe(UnityEngine.Object toDestroy)
+        {
+            if (Application.isPlaying)
+            {
+                Destroy(toDestroy);
+            }
+            else
+            {
+#if UNITY_EDITOR
+                // Defer the destructing in case the object is in the act of being destroyed.
+                EditorApplication.delayCall += () =>
+                {
+                    if (toDestroy != null)
+                    {
+                        DestroyImmediate(toDestroy);
+                    }
+                };
+#endif
+            }
+        }
+    }
+}
