+++ Assets/MixedRealityToolkit/Utilities/Rendering/MaterialInstance.cs.meta
